i guess i'm using firefox print pdf again

+++ 06-echo.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 int main(int argc, char *argv[]) {
-	for (size_t i = 0; i < argc; i++)
+	for (int i = 0; i < argc; i++)
 		printf("%s\n", argv[i]);
 	
 	return EXIT_SUCCESS;
